chore(evidence): split closure module and fix remaining SonarCloud smells
